fontIdentificationTool/requirements_classify.txt
absl-py==1.4.0
asgiref==3.4.1
astor==0.8.1
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.10.1
django-crispy-forms==1.7.2
django-model-utils==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
Flask==2.0.3
Flask-Cors==5.0.0
flatbuffers==1.12
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
idna==3.10
importlib-resources==5.4.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.0.1
oauthlib==3.2.2
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
pdf2image==1.17.0
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==2.4.5
PyPDF2==1.26.0
pytz==2024.2
PyYAML==6.0.1
rdflib==4.2.2
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.22.1
scipy==1.3.2
simplejson==3.19.3
sqlparse==0.4.4
tensorboard==1.15.0
tensorflow==1.15.0
termcolor==1.1.0
tesserocr==2.5.0
uritemplate==4.1.1
urllib3==1.26.20
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.16.0



